--- gaudiondev/next.config.js
@@ -1,11 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  pageExtensions: ['ts', 'tsx', 'js', 'jsx', 'mdx'],
-  experimental: {
-    appDir: true,
-    mdxRs: true,
-  },
-}
-
-const withMDX = require('@next/mdx')()
-module.exports = withMDX(nextConfig)
