+++ gaudiondev/next.config.mjs
@@ -1,0 +1,23 @@
+import addMdx from '@next/mdx';
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  pageExtensions: ['ts', 'tsx', 'js', 'jsx', 'mdx'],
+  experimental: {
+    appDir: true,
+    mdxRs: true,
+  },
+}
+
+
+addMdx(nextConfig, {
+  options: {
+    remarkPlugins: [],
+    rehypePlugins: [],
+    // If you use `MDXProvider`, uncomment the following line.
+    // providerImportSource: "@mdx-js/react",
+  }
+})
+
+
+export default nextConfig;
